部门ID统一为long

+++ src/main/java/cn/felord/domain/callback/CallbackEventBody.java
@@ -70,7 +70,7 @@ public class CallbackEventBody {
     @XStreamAlias("Department")
     private String department;
     @XStreamAlias("MainDepartment")
-    private Integer mainDepartment;
+    private Long mainDepartment;
     @XStreamAlias("IsLeaderInDept")
     private String isLeaderInDept;
     @XStreamAlias("Position")
